+++ scripts/src/common/FormLoader.unit.test.ts
@@ -1,0 +1,5 @@
+import FormLoader from "./FormLoader";
+
+test('dummy test', () => {
+  expect(2).toEqual(2);
+});