+++ webpack.common.js
@@ -52,7 +52,7 @@ module.exports = {
     rules: [
       {
         test: /\.tsx?$/,
-        exclude: /node_modules/,
+        exclude: [/node_modules/, /\.test.tsx?$/],
         use:
           [
             {